t9_plus_ledcontrol/requirements.txt
Flask==3.0.3
requests==2.31.0
pyserial